+++ Sources/LaughPlayer/main.swift
@@ -1,8 +1,9 @@
 import AppKit
 
+LaunchLog.emit("main: starting")
 let app = NSApplication.shared
 let delegate = AppDelegate()
 app.delegate = delegate
 app.setActivationPolicy(.regular)
-app.activate(ignoringOtherApps: true)
+LaunchLog.emit("main: entering run loop")
 app.run()